Final fixes before next iteration
- Changed all measurements back to floating point for more accuracy in logged csv file.
- Adjusted update measurement functions to display shortened floats.
- Adjusted how headers are printed into CSV file (removed spaces)

+++ src/WeatherStation/DisplayToScreen.h
@@ -14,8 +14,8 @@
 void startDisplay(Oled_Screen& display);
 
 //Update respective measurement from WeatherRecord if values change
-void updateTemp(Oled_Screen& display, const int& newTemp);
-void updateHumid(Oled_Screen& display, const int& newHumid);
+void updateTemp(Oled_Screen& display, const float& newTemp);
+void updateHumid(Oled_Screen& display, const float& newHumid);
 void updatePress(Oled_Screen& display, const float& newPress);
 
 //Displaying to console for testing purposes
